requirements.txt
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth==2.35.0
supabase==2.9.1
python-dateutil==2.9.0